+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1251
-    const val appVersion = "4.21.7.1"
+    const val appCodeNumber = 1252
+    const val appVersion = "4.21.7.2"
     const val identifier = "com.unciv.app"
 }